Update README.md
In this example, we have implemented a simple Google Cloud Build trigger that listens for pushes to the main branch of a GitHub repo.

You will need your own Google Cloud project to follow along. Google Cloud offers a free $300 credit for personal accounts, and the free trial can be used to set up and test this.

See the below guides for examples.